% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make.skeleton}
\alias{make.skeleton}
\title{Generate basic files to get an RC package to run}
\usage{
make.skeleton(dir, package.name, short.name, author.name = NULL,
  author.email = NULL)
}
\arguments{
\item{dir}{character Folder where package is being created}

\item{package.name}{character Name of package}

\item{author.name}{character Your name, to place in DESCRIPTION}

\item{author.email}{character Your email, to place in DESCRIPTION}
}
\description{
Generate basic files to get an RC package to run
}

